Vision/requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
requests>=2.31.0
pyyaml>=6.0
pytest>=8.0.0
